pipeline/pyproject.toml
[project]
name = "alphastream-pipeline"
version = "0.1.0"
description = "AlphaStream India - Data Pipeline & ML"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.2",
    "redis[hiredis]>=5.0.3,<6.5",
    "feedparser>=6.0.12",
    "beautifulsoup4>=4.14.3",
    "newspaper4k>=0.9.4",
    "praw>=7.8.1",
    "google-genai>=1.5.0",
    "openai>=2.23.0",
    "spacy>=3.8.11",
    "yfinance>=1.2.0",
    "pandas>=2.2.0",
    "numpy>=2.1.0",
    "scikit-learn>=1.6.0",
    "httpx>=0.28.1",
    "sqlalchemy>=2.0.46,<2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "lxml>=5.0.0",
    "firecrawl-py>=1.0.0",
]

[project.optional-dependencies]
ml = [
    "transformers>=5.2.0",
    "torch>=2.6.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]
